First commit.
I started this project back in early 2014, but back then I didn’t know
git. This is a thinly modified version of the implementation currently
on the SRCF, with no new features, but some files have been
reorganised. In particular, passwords are now stored in a password
file, and cardlists are stored in separate directories according to
their users. Hopefully, this should still be working.